+++ mud-being/src/main/java/com/jpinfo/mudengine/being/model/pk/MudBeingAttrModifierPK.java
@@ -15,7 +15,7 @@ public class MudBeingAttrModifierPK implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Column(name="being_code")
-	private Integer beingCode;
+	private Long beingCode;
 	
 	@Column(name="attr_code")
 	private String attrCode;
@@ -30,10 +30,10 @@ public class MudBeingAttrModifierPK implements Serializable {
 	}
 	
 	
-	public Integer getBeingCode() {
+	public Long getBeingCode() {
 		return this.beingCode;
 	}
-	public void setBeingCode(Integer beingCode) {
+	public void setBeingCode(Long beingCode) {
 		this.beingCode = beingCode;
 	}
 	public String getAttrCode() {
