+++ mud-being/src/main/java/com/jpinfo/mudengine/being/model/pk/MudBeingAttrPK.java
@@ -15,7 +15,7 @@ public class MudBeingAttrPK implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Column(name="being_code", insertable=false, updatable=false)
-	private Integer beingCode;
+	private Long beingCode;
 	
 	@Column(name="attr_code")
 	private String attrCode;
@@ -23,10 +23,10 @@ public class MudBeingAttrPK implements Serializable {
 
 	public MudBeingAttrPK() {
 	}
-	public Integer getBeingCode() {
+	public Long getBeingCode() {
 		return this.beingCode;
 	}
-	public void setBeingCode(Integer beingCode) {
+	public void setBeingCode(Long beingCode) {
 		this.beingCode = beingCode;
 	}
 	public String getAttrCode() {
